removing old implementation of risks
